fix: convert money columns from numeric to bigint (paise)
Fixes all P1 findings from the quality program L3 gate:

Migrations:
- citizen-service/0018: fee.schedules.base_amount, fee.payments.amount,
  fee.refunds.amount → bigint (stores paise)
- legal-service/0021: rti.rti_applications.fee_paid, additional_fee → bigint
- workflow-service/0032: workflow.authority_limits.max_amount → bigint

Schema changes (Drizzle):
- citizen: numeric(14,2) → bigint(mode:'number') on fee payment columns
- legal: numeric(12,2) → bigint(mode:'number') on RTI fee columns
- workflow: numeric(18,2) → bigint(mode:'number') on authority limit amount

Domain logic:
- citizen fee computation: float arithmetic → integer paise math (Math.trunc)
- Validators: z.number() → z.number().int() for all amount inputs
- Removed .toFixed(2) string conversions, pass integers directly

Test fixes:
- Updated rounding assertion (99.999 → 99 with trunc, was 100 with round)
- L3 schema test: excluded fee_pct and credit_hours (not money columns)

All affected service tests pass:
- citizen-service: 378/378 ✅
- legal-service: 306/306 ✅
- workflow-service: 775/775 ✅
- Quality program L3: 38/38 ✅ (was 33/38)

+++ services/workflow-service/src/modules/authority/routes.ts
@@ -32,7 +32,7 @@ export async function authorityRoutes(app: FastifyInstance): Promise<void> {
       scopeRef: z.string().min(1).max(128),
       authorityType: typeEnum.default("financial"),
       currency: z.string().min(1).max(8).default("INR"),
-      maxAmount: z.number().nonnegative(),
+      maxAmount: z.number().int().nonnegative(), // paise
       effectiveFrom: z.string().regex(dateRe),
       effectiveTo: z.string().regex(dateRe).nullable().optional(),
       escalateToScopeType: scopeEnum.nullable().optional(),
